+++ Sources/RouterBytes/Extensions/HTTPRequest+SetBearer.swift
@@ -1,5 +1,5 @@
 //
-//  URLRequest+SetBearer.swift
+//  HTTPRequest+SetBearer.swift
 //  
 //
 //  Created by Lukáš Valenta on 30.04.2023.
@@ -13,13 +13,13 @@ public extension HTTPRequest {
     ///
     /// - Parameter token: A string containing the bearer token to be set in the Authorization header field.
     ///
-    /// - Note: This method mutates the current `URLRequest` instance.
+    /// - Note: This method mutates the current `HTTPRequest` instance.
     ///
     /// - Warning: If the `Authorization` header field is already set, calling this method will overwrite its value.
     ///
     /// Example usage:
     /// ```
-    /// var request = URLRequest(url: URL(string: "https://example.com/api")!)
+    /// var request = HTTPRequest(url: URL(string: "https://example.com/api")!)
     /// request.setBearerToken("abc123")
     /// ```
     @inlinable
@@ -37,7 +37,7 @@ public extension HTTPRequest {
     ///
     /// Example usage:
     /// ```
-    /// var request = URLRequest(url: URL(string: "https://example.com/api")!)
+    /// var request = HTTPRequest(url: URL(string: "https://example.com/api")!)
     /// request.setBearerToken("abc123")
     /// ```
     @inlinable