+++ Tests/RouterBytesTests/APIRouterTests.swift
@@ -56,7 +56,7 @@ final class APIRouterTests: XCTestCase {
         XCTAssertTrue(url.absoluteString.contains("?param1=value1&param2=value2") || url.absoluteString.contains("?param2=value2&param1=value1"))
     }
     
-    func testAsURLRequest() throws {
+    func testAsHTTPRequest() throws {
         let router = BaseAPIRouter<String, Data>(
             hostname: URL(string: "https://example.com")!,
             path: "/api/test",
@@ -96,7 +96,7 @@ final class APIRouterTests: XCTestCase {
         XCTAssertTrue(url.absoluteString.contains("?page=1&perPage=20") || url.absoluteString.contains("?perPage=20&page=1"))
     }
 
-    func testAsURLRequstWithAdditionalHeadersAndQueryItems() throws {
+    func testAsHTTPRequestWithAdditionalHeadersAndQueryItems() throws {
         let router = BaseAPIRouter<EmptyCodable, Data>(
             defaultHeaderFields: [.init("header1")!: "value1", .init("header2")!: "value2"],
             hostname: URL(string: "https://example.com")!,
@@ -117,7 +117,7 @@ final class APIRouterTests: XCTestCase {
         XCTAssertEqual(try router.encodedBody(), try JSONEncoder().encode(EmptyCodable()))
     }
 
-    func testAsURLRequstWithOverridingDefaultHeaderWithAdditionalHeaders() throws {
+    func testAsHTTPRequestWithOverridingDefaultHeaderWithAdditionalHeaders() throws {
         let router = BaseAPIRouter<EmptyCodable, Data>(
             defaultHeaderFields: [.init("Content-Type")!: "test", .init("header1")!: "value1", .init("header2")!: "value2"],
             hostname: URL(string: "https://example.com")!,
@@ -167,7 +167,7 @@ final class APIRouterTests: XCTestCase {
         XCTAssertEqual(request.headerFields[.init("header2")!], "value2")
     }
 
-    func testAsURLRequestWithPOSTMethod() throws {
+    func testAsHTTPRequestWithPOSTMethod() throws {
         struct CreateUserRequest: Encodable {
             let name: String
             let email: String
@@ -187,7 +187,7 @@ final class APIRouterTests: XCTestCase {
         XCTAssertEqual(try router.encodedBody(), try JSONEncoder().encode(createUserRequest))
     }
 
-    func testAsURLRequestWithPUTMethod() throws {
+    func testAsHTTPRequestWithPUTMethod() throws {
         struct UpdateUserRequest: Encodable {
             let id: Int
             let name: String